Assessment_recommendation_System/requirements.txt
streamlit==1.29.0
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
lxml==4.9.3
plotly==5.18.0